requirements.txt
streamlit==1.28.1
pandas==2.1.3
sqlalchemy==2.0.23
requests==2.31.0
beautifulsoup4==4.12.2
google-api-python-client==2.108.0
praw==7.7.1
snscrape==0.7.3.20231007
instagrapi==2.0.0
transformers==4.35.2
torch==2.1.1
vaderSentiment==3.3.2
keybert==0.8.3
matplotlib==3.8.2
wordcloud==1.9.2
plotly==5.17.0
python-dotenv==1.0.0
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1